+++ py/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add `transaction_info` to event payloads, including the transaction's source and internal original transaction name. ([#1330](https://github.com/getsentry/relay/pull/1330))
+
 ## 0.8.13
 
 - Add a data category constant for Replays. ([#1239](https://github.com/getsentry/relay/pull/1239))